sophia: make nsn 64bit

+++ sophia/database/sd_id.h
@@ -14,13 +14,13 @@ typedef struct sdid sdid;
 #define SD_IDBRANCH 1
 
 struct sdid {
-	uint32_t parent;
-	uint32_t id;
+	uint64_t parent;
+	uint64_t id;
 	uint8_t  flags;
 } sspacked;
 
 static inline void
-sd_idinit(sdid *i, uint32_t id, uint32_t parent, uint32_t flags)
+sd_idinit(sdid *i, uint64_t id, uint64_t parent, uint8_t flags)
 {
 	i->id     = id;
 	i->parent = parent;
